+++ Frameworks/BeaconPointProvider.xcframework/ios-arm64/BeaconPointProvider.framework/Headers/PWBeaconPointLocationManager+Provider.h
@@ -1,0 +1,17 @@
+//
+//  PWBeaconPointLocationManager+Provider.h
+//  PWLocation
+//
+//  Created by Troy Stump on 8/4/20.
+//  Copyright © 2020 Phunware. All rights reserved.
+//
+
+#import <PWLocation/PWBeaconPointLocationManager.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface PWBeaconPointLocationManager (Provider) <PWVBLocationManagerProvider>
+
+@end
+
+NS_ASSUME_NONNULL_END